backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
opencv-python-headless
numpy
pydantic
aiofiles
slowapi
python-jose
alembic
pillow
